+++ R/MainFunction.R
@@ -16,8 +16,8 @@
 #'  \cr
 #'  The "\code{classify}" method is not yet implemented.
 #' @param n.iters The number of iterations to run the Gibbs sampler for.
-#' @param range.G Depending on the method employed, either the range of values for the number of clusters, or the conseratively high starting value for the number of clusters. Defaults to 1 for the "\code{FA}" and "\code{IFA}" methods. For the "\code{MFA}" and "\code{MIFA}" models this is to be given as a range of candidate models to explore. For the "\code{OMFA}", "\code{OMIFA}", "\code{IMFA}", and "\code{IMIFA}" models, this is the number of clusters with which the chain is to be initialised, in which case the default is \code{min(N - 1, ceiling(3 * log(N)))}. For the "\code{OMFA}", and "\code{OMIFA}" models this upper limit remains fixed for the entire length of the chain. If \code{length(range.G) * length(range.Q)} is large, consider not storing unnecessary parameters, or breaking up the range of models to be explored into chunks, and send each chunk to \code{\link{get_IMIFA_results}}.
-#' @param range.Q Depending on the method employed, either the range of values for the number of latent factors, or, for methods ending in IFA the conservatively high starting value for the number of cluster-specific factors, in which case the default starting value is \code{floor(3 * log(P))}. For methods ending in IFA, different clusters can be modelled using different numbers of latent factors (incl. zero); for methods not ending in IFA it is possible to fit zero-factor models, corresponding to simple diagonal covariance structures. For instance, fitting the "\code{IMFA}" model with \code{range.Q=0} corresponds to a vanilla Dirichlet Process Mixture Model. If \code{length(range.G) * length(range.Q)} is large, consider not storing unnecessary parameters or breaking up the range of models to be explored into chunks, and send each chunk to \code{\link{get_IMIFA_results}}.
+#' @param range.G Depending on the method employed, either the range of values for the number of clusters, or the conseratively high starting value for the number of clusters. Defaults to 1 for the "\code{FA}" and "\code{IFA}" methods. For the "\code{MFA}" and "\code{MIFA}" models this is to be given as a range of candidate models to explore. For the "\code{OMFA}", "\code{OMIFA}", "\code{IMFA}", and "\code{IMIFA}" models, this is the number of clusters with which the chain is to be initialised, in which case the default is \code{min(N - 1, max(25, ceiling(3 * log(N))))}. For the "\code{OMFA}", and "\code{OMIFA}" models this upper limit remains fixed for the entire length of the chain; \code{range.G} also doubles as the default \code{trunc.G} for the "\code{IMFA}" and "\code{IMIFA}" models. However, when \code{N < P}, or when this bound is close to or exceeds \code{N} for any of these overfitted/infinite mixture models, it is better to initialise at a value closer to the truth (i.e. \code{ceiling(log(N))} by default), though the upper bound remains the same - as a result the role of \code{range.G} when \code{N < P} is no longer to specify the upper bound (which can still be modified via \code{trunc.G}, at least for the "\code{IMFA}" and "\code{IMIFA}" methods) and the number of groups used for initialisation, but rather just the number of groups used for initialisation only. If \code{length(range.G) * length(range.Q)} is large, consider not storing unnecessary parameters, or breaking up the range of models to be explored into chunks, and sending each chunk to \code{\link{get_IMIFA_results}}.
+#' @param range.Q Depending on the method employed, either the range of values for the number of latent factors, or, for methods ending in IFA the conservatively high starting value for the number of cluster-specific factors, in which case the default starting value is \code{floor(3 * log(P))}. For methods ending in IFA, different clusters can be modelled using different numbers of latent factors (incl. zero); for methods not ending in IFA it is possible to fit zero-factor models, corresponding to simple diagonal covariance structures. For instance, fitting the "\code{IMFA}" model with \code{range.Q=0} corresponds to a vanilla Dirichlet Process Mixture Model. If \code{length(range.G) * length(range.Q)} is large, consider not storing unnecessary parameters or breaking up the range of models to be explored into chunks, and sending each chunk to \code{\link{get_IMIFA_results}}.
 #' @param burnin The number of burn-in iterations for the sampler. Defaults to \code{n.iters/5}. Note that chains can also be burned in later, using \code{\link{get_IMIFA_results}}.
 #' @param thinning The thinning interval used in the simulation. Defaults to 2. No thinning corresponds to 1. Note that chains can also be thinned later, using \code{\link{get_IMIFA_results}}.
 #' @param centering A logical value indicating whether mean centering should be applied to the data, defaulting to \code{TRUE}.
@@ -28,8 +28,8 @@
 #' @param alpha Depending on the method employed, either the hyperparameter of the Dirichlet prior for the cluster mixing proportions, or the Dirichlet process concentration parameter. Defaults to 0.5/range.G for the Overfitted methods - if supplied for "\code{OMFA}" and "\code{OMIFA}" methods, you are supplying the numerator of \code{alpha/range.G}, which should be less than half the dimension (per group!) of the free parameters of the smallest model considered in order to ensure superfluous clusters are emptied (for "\code{OMFA}", this corresponds to the smallest \code{range.Q}; for "\code{OMIFA}", this corresponds to a zero-factor model) [see: \code{\link{PGMM_dfree}} and Rousseau and Mengersen (2011)]. Defaults to 1 for the finite mixture models "\code{MFA}" and "\code{MIFA}". Defaults to \code{1 - discount} for the "\code{IMFA}" and "\code{IMIFA}" models if \code{learn.alpha=FALSE} or a simulation from the prior if \code{learn.alpha=TRUE}. Must be positive, unless \code{discount} is supplied for the "\code{IMFA}" or "\code{IMIFA}" methods.
 #' @param psi.alpha The shape of the inverse gamma prior on the uniquenesses. Defaults to 2.5.
 #' @param psi.beta The rate of the inverse gamma prior on the uniquenesses. Can be either a single parameter or a vector of variable specific rates.  If this is not supplied, \code{\link{psi_hyper}} is invoked to choose sensible values, depending on the value of \code{uni.prior}.
-#' @param uni.type A switch indicating whether uniquenesses are to be "\code{unconstrained}" or "\code{isotropic}". Note that unconstrained here means variable-specific and group-specific. Defaults to "\code{unconstrained}" unless \code{N < P}, in which case the default is "\code{isotropic}".
-#' @param uni.prior A switch indicating whether uniquenesses rate hyperparameters are to be "\code{unconstrained}" or "\code{isotropic}". Note that unconstrained here means variable-specific and group-specific. Defaults to \code{uni.type} if that is supplied and \code{uni.prior} is not, otherwise defaults to "\code{unconstrained}" unless \code{N < P}, in which case the default is "\code{isotropic}". Only relevant when "\code{psi.beta}" is not supplied and \code{\link{psi_hyper}} is invoked.
+#' @param uni.type A switch indicating whether uniquenesses are to be "\code{unconstrained}" or "\code{isotropic}". Note that "\code{unconstrained}" here means variable-specific and group-specific, whereas "\code{isotropic}" here means isotropic but still group-specific. The "\code{isotropic}" constraint provides the link between factor analysis and the probabilistic principal component analysis model. Defaults to "\code{unconstrained}", but "\code{isotropic}" is recommended when \code{N < P}.
+#' @param uni.prior A switch indicating whether uniquenesses rate hyperparameters are to be "\code{unconstrained}" or "\code{isotropic}". "\code{uni.prior}" must be "\code{isotropic}" if "\code{uni.type}" is "\code{isotropic}", but can take either value when "\code{uni.type}" is "\code{unconstrained}". Defaults to \code{uni.type} if that is supplied and \code{uni.prior} is not, otherwise defaults to "\code{unconstrained}", but "\code{isotropic}" is recommended when \code{N < P}. Only relevant when "\code{psi.beta}" is not supplied and \code{\link{psi_hyper}} is invoked.
 #' @param z.init The method used to initialise the cluster labels. Defaults to \code{\link[mclust]{Mclust}}. Not relevant for the "\code{FA}" and "\code{"IFA"} methods.
 #' @param z.list A user supplied list of cluster labels. Only relevant if \code{z.init == "z.list"}.
 #' @param adapt A logical value indicating whether adaptation of the number of cluster-specific factors is to take place. Only relevant for methods ending in IFA, in which case the default is \code{TRUE}. Specifying \code{FALSE} and supplying \code{range.Q} provides a means to use the MGP prior in a finite factor context.
@@ -42,9 +42,9 @@
 #' @param nu Hyperparameter for the gamma prior on the local shrinkage parameters. Defaults to 2. Passed to \code{\link{MGP_check}} to ensure validity. Only relevant for methods ending in IFA.
 #' @param nuplus1 Logical switch indicating whether the shape hyperparameter of the prior on the local shrinkage parameters is equal to \code{nu + 1}. If \code{FALSE}, it is simply equal to \code{nu}. Only relevant for methods ending in IFA.
 #' @param adapt.at The iteration at which adaptation is to begin. Defaults to \code{burnin} for the "\code{IFA}" and "\code{MIFA}" methods, defaults to 0 for the "\code{OMIFA}" and "\code{IMIFA}". Cannot exceed \code{burnin}. Only relevant for methods ending in IFA.
-#' @param b0 Intercept parameter for the exponentially decaying adaptation probability s.t. \code{p(iter) = 1/exp(b0 + b1(iter - adapt.at))}. Defaults to 0.1. Only relevant for methods ending in IFA.
-#' @param b1 Slope parameter for the exponentially decaying adaptation probability s.t. \code{p(iter) = 1/exp(b0 + b1(iter - adapt.at))}. Defaults to 0.00005. Only relevant for methods ending in IFA.
-#' @param trunc.G The maximum number of allowable and storable groups if the "\code{IMFA}" or "\code{IMIFA}" method is employed. Defaults to the same value as \code{range.G} and must be greater than or equal to this value. The number of active group to be sampled at each iteration is adaptively truncated, with \code{trunc.G} as an upper limit for storage reasons. Note that large values of \code{trunc.G} may lead to memory capacity issues.
+#' @param b0 Intercept parameter for the exponentially decaying adaptation probability s.t. \code{p(iter) = 1/exp(b0 + b1 * (iter - adapt.at))}. Defaults to 0.1. Only relevant for methods ending in IFA.
+#' @param b1 Slope parameter for the exponentially decaying adaptation probability s.t. \code{p(iter) = 1/exp(b0 + b1 * (iter - adapt.at))}. Defaults to 0.00005. Only relevant for methods ending in IFA.
+#' @param trunc.G The maximum number of allowable and storable groups if the "\code{IMFA}" or "\code{IMIFA}" method is employed. Defaults to the same value as \code{range.G} (unless \code{N < P}, see \code{range.G} for details) and must be greater than or equal to this value. The number of active groups to be sampled at each iteration is adaptively truncated, with \code{trunc.G} as an upper limit for storage reasons. Note that large values of \code{trunc.G} may lead to memory capacity issues.
 #' @param learn.alpha Logical indicating whether the Dirichlet process / Pitman concentration parameter is to be learned, or remain fixed for the duration of the chain. If being learned, a Ga(a, b) prior is assumed for \code{alpha}; updates take place via Gibbs sampling when \code{discount} is zero and via Metropolis-Hastings otherwise. Only relevant for the "\code{IMFA}" and "\code{IMIFA}" methods, in which case the default is \code{TRUE}.
 #' @param alpha.hyper A vector of length 2 giving hyperparameters for the Dirichlet process / Pitman-Yor concentration parameter \code{alpha}. If \code{isTRUE(learn.alpha)}, these are shape and rate parameter of a Gamma distribution. Defaults to Ga(2, 1). Only relevant for the "\code{IMFA}" and "\code{IMIFA}" methods, in which case the default is \code{TRUE}. The prior is shifted to have support on (-\code{discount}, \code{Inf}) when non-zero \code{discount} is supplied or \code{learn.d=TRUE}.
 #' @param zeta Tuning parameter controlling the acceptance rate of the random-walk proposal for the Metropolis-Hastings steps when \code{learn.alpha=TRUE}. These steps are only invoked when either \code{discount} is non-zero or \code{learn.d=TRUE}, otherwise \code{alpha} is learned by Gibbs updates. Must be strictly positive. Defauts to 2.
@@ -80,13 +80,15 @@
 #'
 #' @seealso \code{\link{get_IMIFA_results}}, \code{\link{psi_hyper}}, \code{\link{MGP_check}}
 #' @references
-#' Murphy, K., Gormley, I.C. and Viroli, C. (2017) Infinite Mixtures of Infinite Factor Analysers: Nonparametric Model-Based Clustering via Latent Gaussian Models, \code{https://arxiv.org/abs/1701.07010}
+#' Murphy, K., Gormley, I. C. and Viroli, C. (2017) Infinite Mixtures of Infinite Factor Analysers: Nonparametric Model-Based Clustering via Latent Gaussian Models, \code{https://arxiv.org/abs/1701.07010}
 #'
-#' Bhattacharya, A. and Dunson, D. B. (2011) Sparse Bayesian infinite factor models, \emph{Biometrika}, 98(2): 291–306.
+#' Bhattacharya, A. and Dunson, D. B. (2011) Sparse Bayesian infinite factor models, \emph{Biometrika}, 98(2): 291-306.
 #'
 #' Kalli, M., Griffin, J. E. and Walker, S. G. (2011) Slice sampling mixture models, \emph{Statistics and Computing}, 21(1): 93-105.
 #'
 #' Rousseau, J. and Mengersen, K. (2011) Asymptotic Behaviour of the posterior distribution in overfitted mixture models, \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, 73(5): 689-710.
+#'
+#' Tipping, M. E. and Bishop, C. M. (1999). Probabilistic principal component analysis, \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, 61(3): 611-622.
 #'
 #' @examples
 #' # data(olive)
@@ -100,7 +102,7 @@
 #' # simPY    <- mcmc_IMIFA(olive, method="IMIFA", learn.d=TRUE)
 #' # summary(simPY)
 #'
-#' # Fit a MFA model to the scaled olive data, with isotropic uniquenesses.
+#' # Fit a MFA model to the scaled olive data, with isotropic uniquenesses (i.e. MPPCA).
 #' # Allow diagonal covariance as a special case where range.Q = 0. Accept all other defaults.
 #' # simMFA   <- mcmc_IMIFA(olive, method="MFA", n.iters=10000, range.G=3:6,
 #' #                        range.Q=0:3, centering=FALSE, uni.type="isotropic")
@@ -207,19 +209,20 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
   lnN       <- log(N)
   NlP       <- N < P
   miss.uni  <- missing("uni.type")
-  if(miss.uni)  {
-    uni.type       <- ifelse(NlP, "isotropic", "unconstrained")
-  }
+  if(miss.uni) uni.type      <- "unconstrained"
   if(missing("uni.prior")) {
-    uni.prior      <- ifelse(!miss.uni, uni.type, ifelse(NlP, "isotropic", "unconstrained"))
+    uni.prior      <- uni.type
   }
   uni.type  <- match.arg(uni.type)
   uni.prior <- match.arg(uni.prior)
   if(all(uni.prior == "unconstrained",
          uni.type  == "isotropic")) stop("'uni.prior' can only be 'unconstrained' when 'uni.type' is 'unconstrained'")
+  if(all(uni.prior == "unconstrained",
+         uni.type  == "unconstrained",
+         NlP, miss.uni))            message("Consider setting 'uni.type', or at least 'uni.prior', to 'isotropic' in N << P cases")
 
 # Manage storage switches & warnings for other function inputs
-  if(!missing(mu.switch) && all(!mu.switch, ifelse(method == "classify",
+  if(!missing(mu.switch)  && all(!mu.switch, ifelse(method == "classify",
      !centering, !centered)))       warning("Centering hasn't been applied - are you sure you want mu.switch=FALSE?", call.=FALSE)
   score.x   <- missing(score.switch)
   switches  <- c(mu.sw=mu.switch, s.sw=score.switch, l.sw=load.switch, psi.sw=psi.switch, pi.sw=pi.switch)
@@ -268,10 +271,18 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
     if(all(!G.x, is.element(method, c("FA", "IFA"))) &&
        range.G  > 1)                warning(paste0("'range.G' must be 1 for the ", method, " method"), call.=FALSE)
     if(is.element(method, c("OMIFA", "OMFA", "IMFA", "IMIFA"))) {
-      if(G.x) {
-        range.G    <- as.integer(min(N - 1, max(25, ceiling(3 * log(N)))))
-      }
       lnN2         <- ceiling(lnN)
+      tmp.G        <- as.integer(min(N - 1, max(25, ceiling(3 * lnN))))
+      if(G.x)   {
+        range.G    <- G.init <- tmp.G
+        if(NlP) {                   message(paste0("Since N < P, the sampler will be initialised with a different default of ceiling(log(N)) = ", lnN2, " groups (unless 'range.G' is supplied)"))
+          G.init   <- max(2, lnN2)
+        }
+      }
+      if(all(!G.x, NlP))  {
+        G.init     <- range.G
+        range.G    <- tmp.G
+      }
       if(range.G    < lnN2)         warning(paste0("'range.G' should be at least log(N) (=log(", N, "))", " for the ", method, " method"), call.=FALSE)
       if(is.element(method, c("IMFA", "IMIFA"))) {
         if(any(!is.logical(ind.slice),
@@ -324,7 +335,7 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
       if(length(z.list)    > 1)     stop("Only one set of labels can be supplied via 'z.list'")
       zlabels      <- unlist(z.list)
       if(length(zlabels)  != N)     stop(paste0("'z.list' must be a factor of length N=",  N))
-      if(all(verbose, !missing(range.G)) && any(length(range.G > 1),
+      if(all(verbose, !G.x) && any(length(range.G > 1),
           range.G  != levs))   {    message("Forced 'range.G' equal to the number of levels in 'zlabels' for the 'classify' method")
       }
       range.G      <- levs
@@ -337,15 +348,15 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
     if(G.x)                         stop("'range.G' must be specified")
     if(any(range.G  < 1))           stop("'range.G' must be strictly positive")
     if(any(range.G  > alp3 * lnN))  warning(paste0("'range.G' MUCH greater than log(N) (=log(", N, ")):\n Empty clusters are likely, consider running an overfitted or infinite mixture"), call.=FALSE)
-    range.G <- sort_unique(range.G)
+    range.G <- G.init     <- sort_unique(range.G)
     meth    <- rep(method, length(range.G))
   }
   if(any(range.G >= N))             stop(paste0("'range.G' must be less than the number of observations N=", N))
-  if(range.G[1]  == 1)   {
-    if(is.element(meth[1], c("IMIFA", "IMFA",
-       "OMIFA", "OMFA")))   {       stop("'method' must be FA or IFA for a one group model")
+  if(G.init[1]   == 1)     {
+    if(is.element(method, c("IMIFA", "IMFA",
+       "OMIFA", "OMFA")))  {        stop(paste0("'method' must be FA or IFA for a one group model under the ", method, " method"))
     } else {
-      meth[1]    <- switch(meth[1], MFA=, FA="FA", MIFA=, IFA="IFA")
+      meth[1]    <- switch(method,  MFA=, FA="FA", MIFA=, IFA="IFA")
     }
     if(all(verbose, !is.element(method,
            c("FA", "IFA"))))        message(paste0("Forced use of ", meth[1], " method where 'range.G' is equal to 1"))
@@ -480,14 +491,14 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
   dimension <- PGMM_dfree(Q=switch(method, FA=, classify=, MFA=, OMFA=, IMFA=range.Q,
                IFA=, MIFA=, OMIFA=, IMIFA=0), P=P, method=switch(uni.type, unconstrained="UUU", isotropic="UUC"))
   min.d2    <- min(dimension)/2
-  min.d2G   <- min.d2 * range.G
+  min.d2G   <- min.d2 * G.init
   sw0gs     <- c(mu0g = mu0g, psi0g = psi0g, delta0g = delta0g)
   if(all(!is.element(method, c("MFA", "MIFA", "classify")),
          any(sw0gs)))               stop(paste0("'", names(which(sw0gs)), "' should be FALSE for the ", method, " method\n"))
   if(!is.element(method, c("FA", "IFA", "classify"))) {
-    if(missing("alpha"))   { alpha         <- switch(method, MFA=, MIFA=1, OMFA=, OMIFA=0.5/range.G, if(learn.alpha) max(1, rgamma(1, a.hyp1, a.hyp2)) - discount else 1 - discount)
+    if(missing("alpha"))   { alpha         <- switch(method, MFA=, MIFA=1, OMFA=, OMIFA=0.5/G.init, if(learn.alpha) max(1, rgamma(1, a.hyp1, a.hyp2)) - discount else 1 - discount)
     } else if(is.element(method,
-      c("OMFA", "OMIFA")))   alpha         <- alpha/range.G
+      c("OMFA", "OMIFA")))   alpha         <- alpha/G.init
     if(length(alpha) != 1)          stop("'alpha' must be specified as a scalar to ensure an exchangeable prior")
     if(alpha <= -discount)          stop(paste0("'alpha' must be ",     ifelse(discount != 0, paste0("strictly greater than -discount (i.e. > ", - discount, ")"), "strictly positive")))
     if(all(is.element(method,  c("IMIFA",   "IMFA")),
@@ -496,7 +507,7 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
        alpha >= min.d2))            warning(paste0("'alpha' over 'range.G' for the OMFA & OMIFA methods must be less than half the dimension (per group!)\n of the free parameters of the smallest model considered (= ", min.d2, "): consider suppling 'alpha' < ", min.d2G), call.=FALSE)
     if(any(all(is.element(method, c("MFA",  "MIFA")), alpha > 1),
            all(is.element(method, c("OMFA", "OMIFA")),
-           alpha > 1/range.G)))     warning("Are you sure alpha should be greater than 1?", call.=FALSE)
+           alpha > 1/G.init)))      warning("Are you sure alpha should be greater than 1?", call.=FALSE)
                              z.init        <- match.arg(z.init)
     if(all(any(is.nan(rDirichlet(G=min(range.G), alpha = alpha))),
            is.element(method,  c("MFA",     "MIFA",
@@ -507,11 +518,11 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
       if(length(z.list)   != len.G)  {
                                     stop(paste0("'z.list' must be a list of length ", len.G))  }
                              list.levels   <- lapply(z.list, nlevels)
-      if(!all(list.levels == range.G))            {
+      if(!all(list.levels == G.init))             {
         if(!is.element(method, c("IMIFA",
                        "IMFA", "OMIFA", "OMFA"))) {
                                     stop(paste0("Each element of 'z.list' must have the same number of levels as 'range.G'"))
-        } else                      stop(paste0("Only ", list.levels, " groups are populated according to z.list, but 'range.G' has been set to ", range.G, ":\n  Reset range.G to this value to avoid redunandtly carrying around empty groups or supply a list with ", range.G, " levels"))
+        } else                      stop(paste0("Only ", list.levels, " groups are populated according to z.list, but 'range.G' has been set to ", G.init, ":\n  Reset 'range.G' to this value to avoid redunandtly carrying around empty groups or supply a list with ", G.init, " levels"))
       }
       if(!all(lengths(z.list) == N)) {
                                     stop(paste0("Each element of 'z.list' must be a vector of length N=", N)) }
@@ -552,12 +563,12 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
   if(beta.x) {
     psi.beta       <- temp.psi <- list(psi_hyper(shape=psi.alpha, covar=cov.mat, type=uni.prior))
   } else {
-    psi.beta       <- .len_check(psi.beta, psi0g, method, P, range.G)
+    psi.beta       <- .len_check(psi.beta, psi0g, method, P, G.init)
   }
-  mu.zero          <- if(mu0.x) mu else .len_check(mu.zero, mu0g, method, P, range.G)
+  mu.zero          <- if(mu0.x) mu else .len_check(mu.zero, mu0g, method, P, G.init)
   if(!is.element(method, c("FA", "MFA", "OMFA", "IMFA"))) {
-    alpha.d1       <- if(ad1.x) list(3L) else .len_check(alpha.d1, delta0g, method, P, range.G, P.dim=FALSE)
-    alpha.d2       <- if(ad2.x) list(6L) else .len_check(alpha.d2, delta0g, method, P, range.G, P.dim=FALSE)
+    alpha.d1       <- if(ad1.x) list(3L) else .len_check(alpha.d1, delta0g, method, P, G.init, P.dim=FALSE)
+    alpha.d2       <- if(ad2.x) list(6L) else .len_check(alpha.d2, delta0g, method, P, G.init, P.dim=FALSE)
     if(all(NlP, any(ad1.x, ad2.x))) message("Consider applying more shrinkage with higher 'alpha.d1' and 'alpha.d2' hyperparameter values when N << P")
   }
   if(!is.element(method, c("FA", "IFA"))) {
@@ -566,9 +577,9 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
     pi.alpha       <- list()
     pi.prop        <- list()
     zi             <- list()
-    for(g in seq_along(range.G)) {
-      G            <- range.G[g]
-      if(z.init    == "kmeans")     {
+    for(g in seq_along(G.init)) {
+      G            <- G.init[g]
+      if(z.init    == "kmeans")      {
         k.res      <- try(suppressWarnings(kmeans(dat, centers=G, iter.max=20, nstart=100)), silent=TRUE)
         if(!inherits(k.res, "try-error"))  {
           zi[[g]]  <- as.integer(factor(k.res$cluster,        levels=seq_len(G)))
@@ -589,8 +600,8 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
             zips   <- .sim_z_p(N=N, prob.z=pies)
           }
         } else {
-          vies     <- .sim_vs_inf(alpha=alpha, N=N, nn=rep(N/range.G, range.G), discount=discount, len=range.G, lseq=seq_len(range.G))
-          pies     <- .sim_pi_inf(vs=vies, len=range.G)
+          vies     <- .sim_vs_inf(alpha=alpha, N=N, nn=rep(N/G.init, G.init), discount=discount, len=G.init, lseq=seq_len(G.init))
+          pies     <- .sim_pi_inf(vs=vies, len=G.init)
           zips     <- .sim_z_p(N=N, prob.z=pies)
         }
         zi[[g]]    <- as.integer(zips)
@@ -653,10 +664,11 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
   if(is.element(method, c("classify", "IFA", "MIFA", "IMIFA", "OMIFA"))) {
     if(!all(MGP_check(ad1=unlist(alpha.d1), ad2=unlist(alpha.d2), Q=unique(range.Q), nu=nu, bd1=beta.d1, bd2=beta.d2,
         plus1=nuplus1)[[1]]$valid)) stop("Invalid shrinkage hyperparameter values will not encourage loadings column removal.\n Try using the MGP_check() function in advance to ensure cumulative shrinkage property holds.")
-    deltas         <- lapply(seq_along(range.G), function(g) list(alpha.d1 = alpha.d1[[g]], alpha.d2 = alpha.d2[[g]]))
+    deltas         <- lapply(seq_along(G.init), function(g) list(alpha.d1 = alpha.d1[[g]], alpha.d2 = alpha.d2[[g]]))
   }
   init.time        <- proc.time() - init.start
   fac.time         <- 0
+  G.range          <- switch(method, IMIFA=, IMFA=G.init, range.G)
 
   if(is.element(method, c("IFA", "MIFA", "OMIFA", "IMIFA"))) {
     if(len.G == 1)  {
@@ -665,13 +677,13 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
         gibbs.arg  <- append(temp.args, deltas[[Gi]])
       }
       imifa[[Gi]][[Qi]]   <- do.call(paste0(".gibbs_", meth[Gi]),
-                                     args=append(list(data = dat, N = N, G = range.G, Q = range.Q, mu = mu[[Gi]], mu.zero = mu.zero[[Gi]],
+                                     args=append(list(data = dat, N = N, G = G.range, Q = range.Q, mu = mu[[Gi]], mu.zero = mu.zero[[Gi]],
                                                       psi.beta = psi.beta[[Gi]], cluster = if(meth[Gi] != "IFA") clust[[Gi]]), gibbs.arg))
       fac.time     <- fac.time + imifa[[Gi]][[Qi]]$time
     } else {
       start.time   <- proc.time()
-      for(g in range.G) {
-        Gi         <- which(range.G == g)
+      for(g in G.range)   {
+        Gi         <- which(G.range == g)
         if(meth[Gi]  == "IFA") {
          gibbs.arg <- append(temp.args, deltas[[Gi]])
         }
@@ -687,23 +699,23 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
     if(all(len.G == 1, len.Q == 1)) {
       start.time   <- proc.time()
       imifa[[Gi]][[Qi]]   <- do.call(paste0(".gibbs_", meth[Gi]),
-                                     args=append(list(data = dat, N = N, G = range.G, Q = range.Q, mu = mu[[Gi]], mu.zero = mu.zero[[Gi]],
-                                                      psi.beta = psi.beta[[Gi]], cluster = if(meth[Gi] != "FA") clust[[Gi]]), gibbs.arg))
+                                     args=append(list(data = dat, N = N, G = G.range, Q = range.Q, mu = mu[[Gi]], mu.zero = mu.zero[[Gi]],
+                                                      psi.beta = psi.beta[[Gi]], cluster = if(meth[Gi] !=  "FA") clust[[Gi]]), gibbs.arg))
       fac.time     <- fac.time + imifa[[Gi]][[Qi]]$time
     } else if(len.G == 1) {
       start.time   <- proc.time()
       for(q in range.Q)   {
         Qi         <- which(range.Q == q)
         imifa[[Gi]][[Qi]] <- do.call(paste0(".gibbs_", meth[Gi]),
-                                     args=append(list(data = dat, N = N, G = range.G, Q = q, mu = mu[[Gi]], mu.zero = mu.zero[[Gi]],
-                                                      psi.beta = psi.beta[[Gi]], cluster = if(meth[Gi] != "FA") clust[[Gi]]), gibbs.arg))
+                                     args=append(list(data = dat, N = N, G = G.range, Q = q, mu = mu[[Gi]], mu.zero = mu.zero[[Gi]],
+                                                      psi.beta = psi.beta[[Gi]], cluster = if(meth[Gi] !=  "FA") clust[[Gi]]), gibbs.arg))
         fac.time   <- fac.time + imifa[[Gi]][[Qi]]$time
         if(verbose && Qi  != len.Q) cat(paste0("Model ", Qi, " of ", len.Q, " complete"), "Initialising...", sep="\n")
       }
     } else if(len.Q == 1) {
       start.time   <- proc.time()
-      for(g in range.G) {
-        Gi         <- which(range.G == g)
+      for(g in G.range)   {
+        Gi         <- which(G.range == g)
         imifa[[Gi]]       <- list()
         imifa[[Gi]][[Qi]] <- do.call(paste0(".gibbs_", meth[Gi]),
                                      args=append(list(data = dat, N = N, G = g, Q = range.Q, mu = mu[[Gi]], mu.zero = mu.zero[[Gi]],
@@ -714,8 +726,8 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
     } else {
       mi           <- 0
       start.time   <- proc.time()
-      for(g in range.G) {
-        Gi         <- which(range.G == g)
+      for(g in G.range)   {
+        Gi         <- which(G.range == g)
         imifa[[Gi]]       <- list()
         for(q in range.Q) {
           Qi       <- which(range.Q == q)
@@ -760,12 +772,12 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
      lapply(seq_along(imifa), function(x) setNames(imifa[[x]], paste0(range.Q, ifelse(range.Q == 1, "Factor", "Factors"))))
   }, lapply(seq_along(imifa), function(x) setNames(imifa[[x]], "IFA")))
   if(!is.element(method, c("FA", "IFA"))) {
-    for(g in seq_along(range.G)) {
+    for(g in seq_along(G.init)) {
       attr(imifa[[g]],
-           "Z.init")      <- factor(zi[[g]], levels=seq_len(range.G[g]))
+           "Z.init")      <- factor(zi[[g]], levels=seq_len(G.init[g]))
     }
   }
-  gnames    <- switch(method, classify=paste0("Group ", seq_len(range.G)), paste0(range.G, ifelse(range.G == 1, "Group", "Groups")))
+  gnames    <- switch(method, classify=paste0("Group ", seq_len(range.G)), paste0(G.init, ifelse(G.init == 1, "Group", "Groups")))
   names(imifa)            <- gnames
   attr(imifa,
        "Alph.step")       <- learn.alpha
@@ -814,6 +826,8 @@ mcmc_IMIFA  <- function(dat = NULL, method = c("IMIFA", "IMFA", "OMIFA", "OMFA",
   }
   if(is.element(method, c("FA", "IFA", "classify"))) {
     times                 <- times[-length(times)]
+  } else {
+    attr(imifa, "G.init") <- G.init
   }
   class(times)            <- "listof"
   attr(imifa, "Time")     <- times